[Correção] Tipagem de vars no VS 2015.
Correções referentes a alertas emitidos durante o processo de compilação
no VS 2015.

+++ src/map/clif.c
@@ -1919,7 +1919,7 @@ void clif_selllist(struct map_session_data *sd)
 /// - append this text
 void clif_scriptmes(struct map_session_data *sd, int npcid, const char *mes) {
 	int fd = sd->fd;
-	size_t slen;
+	unsigned int slen;
 
 	nullpo_retv(sd);
 	nullpo_retv(mes);
@@ -2034,7 +2034,7 @@ void clif_sendfakenpc(struct map_session_data *sd, int npcid) {
 /// TODO investigate behavior of other windows [FlavioJS]
 void clif_scriptmenu(struct map_session_data* sd, int npcid, const char* mes) {
 	int fd;
-	size_t slen;
+	unsigned int slen;
 	struct block_list *bl = NULL;
 
 	nullpo_retv(sd);
@@ -3571,7 +3571,7 @@ void clif_dispchat(struct chat_data* cd, int fd)
 	     : 1;
 
 	WBUFW(buf, 0) = 0xd7;
-	WBUFW(buf, 2) = 17 + strlen(cd->title);
+	WBUFW(buf, 2) = 17 + (uint16)strlen(cd->title);
 	WBUFL(buf, 4) = cd->owner->id;
 	WBUFL(buf, 8) = cd->bl.id;
 	WBUFW(buf,12) = cd->limit;
@@ -3608,7 +3608,7 @@ void clif_changechatstatus(struct chat_data* cd)
 	     : 1;
 
 	WBUFW(buf, 0) = 0xdf;
-	WBUFW(buf, 2) = 17 + strlen(cd->title);
+	WBUFW(buf, 2) = 17 + (uint16)strlen(cd->title);
 	WBUFL(buf, 4) = cd->owner->id;
 	WBUFL(buf, 8) = cd->bl.id;
 	WBUFW(buf,12) = cd->limit;
@@ -5535,7 +5535,7 @@ void clif_displaymessage2(const int fd, const char* mes) {
 		line = strtok(message, "\n");
 		while(line != NULL) {
 			// Limit message to 255+1 characters (otherwise it causes a buffer overflow in the client)
-			size_t len = strnlen(line, 255);
+			unsigned int len = strnlen(line, 255);
 
 			if (len > 0) { // don't send a void message (it's not displaying on the client chat). @help can send void line.
 				if( map->cpsd_active && fd == 0 ) {
@@ -5589,7 +5589,7 @@ void clif_displaymessage_sprintf(const int fd, const char *mes, ...) {
 }
 /// Send broadcast message in yellow or blue without font formatting (ZC_BROADCAST).
 /// 009a <packet len>.W <message>.?B
-void clif_broadcast(struct block_list *bl, const char *mes, size_t len, int type, enum send_target target)
+void clif_broadcast(struct block_list *bl, const char *mes, uint32_t len, int type, enum send_target target)
 {
 	int lp = (type&BC_COLOR_MASK) ? 4 : 0;
 	unsigned char *buf = NULL;
@@ -5615,7 +5615,7 @@ void clif_broadcast(struct block_list *bl, const char *mes, size_t len, int type
  *------------------------------------------*/
 void clif_GlobalMessage(struct block_list* bl, const char* message) {
 	char buf[256];
-	size_t len;
+	unsigned int len;
 	nullpo_retv(bl);
 
 	if(!message)
@@ -5638,7 +5638,7 @@ void clif_GlobalMessage(struct block_list* bl, const char* message) {
 
 /// Send broadcast message with font formatting (ZC_BROADCAST2).
 /// 01c3 <packet len>.W <fontColor>.L <fontType>.W <fontSize>.W <fontAlign>.W <fontY>.W <message>.?B
-void clif_broadcast2(struct block_list* bl, const char* mes, size_t len, unsigned int fontColor, short fontType, short fontSize, short fontAlign, short fontY, enum send_target target)
+void clif_broadcast2(struct block_list* bl, const char* mes, uint32_t len, unsigned int fontColor, short fontType, short fontSize, short fontAlign, short fontY, enum send_target target)
 {
 	unsigned char *buf;
 
@@ -5810,7 +5810,7 @@ void clif_upgrademessage(int fd, int result, int item_id)
 /// Whisper is transmitted to the destination player (ZC_WHISPER).
 /// 0097 <packet len>.W <nick>.24B <message>.?B
 /// 0097 <packet len>.W <nick>.24B <isAdmin>.L <message>.?B (PACKETVER >= 20091104)
-void clif_wis_message(int fd, const char *nick, const char *mes, size_t mes_len)
+void clif_wis_message(int fd, const char *nick, const char *mes, unsigned int mes_len)
 {
 #if PACKETVER >= 20091104
 	struct map_session_data *ssd = NULL;
@@ -7953,7 +7953,7 @@ void clif_marriage_proposal(int fd, struct map_session_data *sd, struct map_sess
 /*==========================================
  * Displays a message using the guild-chat colors to the specified targets. [Skotlex]
  *------------------------------------------*/
-void clif_disp_message(struct block_list* src, const char* mes, size_t len, enum send_target target)
+void clif_disp_message(struct block_list* src, const char* mes, uint32_t len, enum send_target target)
 {
 	unsigned char buf[256];
 
@@ -8219,7 +8219,7 @@ void clif_specialeffect_value(struct block_list* bl, int effect_id, int num, sen
  */
 void clif_messagecolor_self(int fd, uint32 color, const char *msg)
 {
-	size_t msg_len;
+	unsigned int msg_len;
 
 	nullpo_retv(msg);
 	msg_len = strlen(msg) + 1;
@@ -8244,7 +8244,7 @@ void clif_messagecolor_self(int fd, uint32 color, const char *msg)
  */
 void clif_messagecolor(struct block_list* bl, uint32 color, const char *msg)
 {
-	size_t msg_len = strlen(msg) + 1;
+	unsigned int msg_len = strlen(msg) + 1;
 	uint8 buf[256];
 
 	nullpo_retv(bl);
@@ -8577,7 +8577,7 @@ void clif_slide(struct block_list *bl, int x, int y)
 void clif_disp_overhead(struct block_list *bl, const char* mes)
 {
 	unsigned char buf[256]; //This should be more than sufficient, the theoretical max is CHAT_SIZE + 8 (pads and extra inserted crap)
-	size_t len_mes;
+	unsigned int len_mes;
 
 	nullpo_retv(bl);
 	nullpo_retv(mes);
@@ -8990,7 +8990,7 @@ void clif_channel_msg(struct channel_data *chan, struct map_session_data *sd, ch
 {
 	DBIterator *iter;
 	struct map_session_data *user;
-	unsigned short msg_len;
+	unsigned int msg_len;
 	uint32 color;
 
 	nullpo_retv(chan);
@@ -9025,7 +9025,7 @@ void clif_channel_msg2(struct channel_data *chan, char *msg)
 	DBIterator *iter;
 	struct map_session_data *user;
 	unsigned char buf[210];
-	unsigned short msg_len;
+	unsigned int msg_len;
 	uint32 color;
 
 	nullpo_retv(chan);
@@ -9735,7 +9735,7 @@ void clif_parse_GlobalMessage(int fd, struct map_session_data* sd) __attribute__
 void clif_parse_GlobalMessage(int fd, struct map_session_data* sd)
 {
 	const char* text = (char*)RFIFOP(fd,4);
-	size_t textlen = RFIFOW(fd,2) - 4;
+	unsigned int textlen = RFIFOW(fd,2) - 4;
 
 	char *name, *message, *fakename = NULL;
 	size_t namelen, messagelen;
@@ -14645,7 +14645,7 @@ void clif_Mail_read(struct map_session_data *sd, int mail_id)
 		struct mail_message *msg = &sd->mail.inbox.msg[i];
 		struct item *item = &msg->item;
 		struct item_data *data;
-		size_t msg_len = strlen(msg->body), len;
+		unsigned int msg_len = strlen(msg->body), len;
 
 		if( msg_len == 0 ) {
 			strcpy(msg->body, "(nenhuma mensagem)");
@@ -16072,7 +16072,7 @@ void clif_bg_xy_remove(struct map_session_data *sd)
 
 /// Notifies clients of a battleground message (ZC_BATTLEFIELD_CHAT).
 /// 02dc <packet len>.W <account id>.L <name>.24B <message>.?B
-void clif_bg_message(struct battleground_data *bgd, int src_id, const char *name, const char *mes, size_t len)
+void clif_bg_message(struct battleground_data *bgd, int src_id, const char *name, const char *mes, unsigned int len)
 {
 	struct map_session_data *sd;
 	unsigned char *buf;
@@ -17608,7 +17608,7 @@ void clif_partytickack(struct map_session_data* sd, bool flag) {
 
 void clif_ShowScript(struct block_list* bl, const char* message) {
 	char buf[256];
-	size_t len;
+	uint32_t len;
 	nullpo_retv(bl);
 
 	if(!message)
