+++ src/map/clif.h
@@ -833,9 +833,9 @@ struct clif_interface {
 	void (*changechatstatus) (struct chat_data* cd);
 	void (*wis_message) (int fd, const char* nick, const char* mes, size_t mes_len);
 	void (*wis_end) (int fd, int flag);
-	void (*disp_message) (struct block_list* src, const char* mes, size_t len, enum send_target target);
-	void (*broadcast) (struct block_list* bl, const char* mes, size_t len, int type, enum send_target target);
-	void (*broadcast2) (struct block_list* bl, const char* mes, size_t len, unsigned int fontColor, short fontType, short fontSize, short fontAlign, short fontY, enum send_target target);
+	void (*disp_message) (struct block_list* src, const char* mes, uint32_t len, enum send_target target);
+	void (*broadcast) (struct block_list* bl, const char* mes, uint32_t len, int type, enum send_target target);
+	void (*broadcast2) (struct block_list* bl, const char* mes, uint32_t len, unsigned int fontColor, short fontType, short fontSize, short fontAlign, short fontY, enum send_target target);
 	void (*messagecolor_self) (int fd, uint32 color, const char *msg);
 	void (*messagecolor) (struct block_list* bl, uint32 color, const char* msg);
 	void (*disp_overhead) (struct block_list *bl, const char* mes);
